Moving components to appropriate places
