+++ apps/legal-gazette-application-web/src/app/(application-shell)/@footer/umsoknir/innkollun-throtabus/[id]/page.tsx
@@ -1,12 +1,12 @@
 'use client'
 
-import { BankruptcyFooter } from '../../../../../../components/client-components/application/bankruptcy/BankruptcyFooter'
 import { ApplicationFooter } from '../../../../../../components/client-components/application/footer/ApplicationFooter'
+import { BankruptcyFormFooter } from '../../../../../../components/client-components/form/bankruptcy/BankruptcyFormFooter'
 
-export default function ThrotabusFooter() {
+export default function ThrotabusFooter(props: any) {
   return (
     <ApplicationFooter>
-      <BankruptcyFooter />
+      <BankruptcyFormFooter />
     </ApplicationFooter>
   )
 }
