+++ apps/legal-gazette-application-web/src/app/(application-shell)/umsoknir/innkollun-throtabus/[id]/page.tsx
@@ -3,7 +3,7 @@ import { getServerSession } from 'next-auth'
 
 import { getLogger } from '@dmr.is/logging'
 
-import { BankruptcyForm } from '../../../../../components/client-components/application/bankruptcy/BankruptcyForm'
+import { BankruptcyForm } from '../../../../../components/client-components/form/bankruptcy/BankruptcyForm'
 import { BankruptcyCase } from '../../../../../components/client-components/case/BankruptcyCase'
 import { BankruptcyApplicationDtoStatusEnum } from '../../../../../gen/fetch'
 import { authOptions } from '../../../../../lib/authOptions'
